refactor: remove dead code, trim models/config/tests (~1700 lines)
