+++ config.yaml
@@ -16,18 +16,10 @@
 # paths where documented in settings.py.
 # =============================================================================
 
-ai_features:
-  enable_ab_testing: true
-  enable_advanced_audio: true
-  enable_auto_optimization: true
-  enable_cinematic_editing: true
-  enable_proactive_management: true
-  fallback_confidence_threshold: 0.6
-  use_fallback_analysis: true
 api:
   nvidia_nim_api_key: ""
   nvidia_nim_model: "moonshotai/kimi-k2.6"
-  nvidia_nim_alt_model: "qwen/qwen3-235b-a22b"
+  nvidia_nim_alt_model: "minimaxai/minimax-m3"
   nvidia_nim_base_url: "https://integrate.api.nvidia.com/v1"
   nvidia_nim_rate_limit_rpm: 60
   ai_provider: "nvidia_nim"
@@ -87,133 +79,12 @@ audio:
     enable: true
     normalize_levels: true
     volume: 0.7
-audio_processing:
-  clarity_boost: 0.2
-  compression_ratio: 0.4
-  duck_during_narration: true
-  duck_volume: 0.3
-  enable_smart_ducking: true
-  enable_voice_enhancement: true
-  fade_duration: 0.5
-  music_analysis_window: 2048
-  noise_reduction_strength: 0.3
-  preserve_music_dynamics: true
-  sample_rate: 44100
-  smart_detection: true
-  voice_frequency_range:
-  - 80
-  - 8000
-channel_management:
-  auto_heart_positive_comments: true
-  auto_pin_high_value_comments: true
-  auto_respond_to_comments: true
-  comment_analysis_interval: 300
-  enable_ai_responses: true
-  enable_proactive_management: true
-  heart_comment_threshold: 60
-  high_engagement_threshold: 80
-  max_comments_per_analysis: 50
-  pin_comment_threshold: 85
-  response_templates:
-    appreciation: "Really appreciate the positive feedback! \U0001F60A"
-    general: "Thanks for watching and commenting! \U0001F44D"
-    positive: "Thank you so much! Glad you enjoyed the content! \U0001F64F"
-    question: Great question! Let me think about doing a video on this topic!
-  toxicity_threshold: 20
-cinematic_editing:
-  composition_analysis: true
-  contrast_weight: 0.3
-  enable_cinematic_transitions: true
-  enable_dynamic_movements: true
-  leading_lines_weight: 0.2
-  max_zoom_factor: 1.8
-  min_scene_duration: 2.0
-  motion_threshold: 0.3
-  movement_intensity_range:
-  - 0.3
-  - 1.5
-  rule_of_thirds_weight: 0.3
-  symmetry_weight: 0.2
-  transition_duration_range:
-  - 0.2
-  - 0.5
-  visual_complexity_weight: 0.4
 database:
   ab_test_retention_days: 30
   cleanup_interval_hours: 24
   metrics_retention_days: 90
   optimization_history_retention: 50
   sqlite_db_path: data/youtube_shorts.db
-development:
-  debug_mode: false
-  enable_profiling: false
-  mock_api_calls: false
-  save_intermediate_files: false
-  skip_upload_in_test_mode: true
-  test_video_duration_limit: 60
-  use_sample_data: false
-feature_flags:
-  enable_advanced_analytics: true
-  enable_auto_captions: true
-  enable_auto_publishing: false
-  enable_experimental_ai: false
-  enable_real_time_monitoring: true
-  enable_style_transfer: false
-  enable_voice_cloning: false
-logging:
-  backup_count: 5
-  component_levels:
-    ai_client: DEBUG
-    audio_processor: INFO
-    channel_manager: INFO
-    cinematic_editor: INFO
-    enhancement_optimizer: INFO
-    thumbnail_generator: INFO
-  console_logging: true
-  file_logging: true
-  format: '%(asctime)s - %(name)s - %(levelname)s - %(message)s'
-  level: INFO
-  log_rotation: true
-  max_log_size_mb: 10
-optimization:
-  confidence_threshold: 0.7
-  enable_auto_optimization: true
-  max_adjustment_per_cycle: 0.2
-  min_sample_size: 10
-  optimization_cycle_days: 7
-  parameter_constraints:
-    background_music_volume:
-      max: 0.8
-      min: 0.1
-      step: 0.05
-    color_grading_strength:
-      max: 1.0
-      min: 0.0
-      step: 0.05
-    hook_text_font_size:
-      max: 120
-      min: 40
-      step: 5
-    sound_effects_volume:
-      max: 1.0
-      min: 0.1
-      step: 0.05
-    speed_effect_factor:
-      max: 2.0
-      min: 0.5
-      step: 0.1
-    text_overlay_duration:
-      max: 5.0
-      min: 1.0
-      step: 0.2
-    visual_effects_intensity:
-      max: 2.0
-      min: 0.3
-      step: 0.1
-    zoom_intensity:
-      max: 1.8
-      min: 1.0
-      step: 0.1
 paths:
   ab_test_results_dir: ab_test_results
   ai_models_cache_dir: ai_models_cache
@@ -228,101 +99,8 @@ paths:
   sound_effects_dir: sound_effects
   temp_dir: temp
   thumbnails_dir: thumbnails
-performance:
-  aggressive_memory_cleanup: true
-  chunk_size_mb: 100
-  clear_cache_between_videos: true
-  enable_gpu_acceleration: true
-  gpu_memory_management: true
-  max_concurrent_videos: 3
-  max_vram_usage: 0.7
-  memory_monitoring_interval: 30
-thumbnail_optimization:
-  ab_test_variants: 3
-  color_schemes:
-    classic_bold:
-      bg_overlay: 0.1
-      stroke_color: black
-      text_color: white
-    cool_dramatic:
-      bg_overlay: 0.4
-      stroke_color: blue
-      text_color: cyan
-    high_contrast:
-      bg_overlay: 0.3
-      stroke_color: black
-      text_color: white
-    neon_pop:
-      bg_overlay: 0.25
-      stroke_color: magenta
-      text_color: lime
-    warm_bright:
-      bg_overlay: 0.2
-      stroke_color: red
-      text_color: yellow
-  emotional_tones:
-    dramatic:
-      brightness: 0.9
-      contrast: 1.4
-      saturation: 0.9
-    energetic:
-      brightness: 1.2
-      contrast: 1.1
-      saturation: 1.4
-    exciting:
-      brightness: 1.1
-      contrast: 1.2
-      saturation: 1.3
-    mysterious:
-      brightness: 0.8
-      contrast: 1.3
-      saturation: 0.7
-    professional:
-      brightness: 1.0
-      contrast: 1.1
-      saturation: 1.0
-  enable_ab_testing: true
-  max_variants: 5
-  min_impressions_for_test: 1000
-  performance_window_days: 7
-  test_duration_hours: 24
-  text_styles:
-    bold_impact:
-      font_weight: bold
-      size_factor: 1.2
-      spacing: 1.1
-    compact_punch:
-      font_weight: bold
-      size_factor: 0.9
-      spacing: 0.9
-    dramatic_tall:
-      font_weight: bold
-      size_factor: 1.0
-      spacing: 1.3
-    elegant_modern:
-      font_weight: normal
-      size_factor: 1.1
-      spacing: 1.0
-    explosive_caps:
-      font_weight: bold
-      size_factor: 1.3
-      spacing: 1.2
 tts:
   narrator_speaker: Ryan
-  azure:
-    pitch: medium
-    speaking_rate: medium
-    voice_name: en-US-AriaNeural
-  edge_tts:
-    pitch: +0Hz
-    rate: +0%
-    voice: en-US-AriaNeural
-  elevenlabs:
-    enable_voice_enhancement: true
-    model_id: eleven_monolingual_v1
-    voice_id: 21m00Tcm4TlvDq8ikWAM
-    voice_similarity_boost: 0.75
-    voice_stability: 0.75
   primary_service: elevenlabs
 video_processing:
   audio_bitrate: 128k
