+++ src/database/db_manager.py
@@ -67,10 +67,7 @@ class DatabaseManager:
 
                 conn.commit()
 
-                # Check and perform migrations
-                self._migrate_database(cursor)
-
-                # Create indexes for performance (after migrations)
+                # Create indexes for performance
                 cursor.execute(
                     "CREATE INDEX IF NOT EXISTS idx_uploads_reddit_url ON uploads(reddit_url)"
                 )
@@ -100,146 +97,6 @@ class DatabaseManager:
         except sqlite3.Error as e:
             self.logger.error(f"Database initialization failed: {e}")
             raise
-
-    def _migrate_database(self, cursor: sqlite3.Cursor):
-        """Perform database migrations for schema updates"""
-        try:
-            # Get current table schema
-            cursor.execute("PRAGMA table_info(uploads)")
-            columns = {row[1] for row in cursor.fetchall()}
-
-            # Add missing columns from older versions
-            migrations = [
-                (
-                    "reddit_post_id",
-                    "ALTER TABLE uploads ADD COLUMN reddit_post_id TEXT",
-                ),
-                (
-                    "youtube_video_id",
-                    "ALTER TABLE uploads ADD COLUMN youtube_video_id TEXT",
-                ),
-                (
-                    "original_score",
-                    "ALTER TABLE uploads ADD COLUMN original_score INTEGER",
-                ),
-                (
-                    "processing_duration_seconds",
-                    "ALTER TABLE uploads ADD COLUMN processing_duration_seconds REAL",
-                ),
-                (
-                    "video_duration_seconds",
-                    "ALTER TABLE uploads ADD COLUMN video_duration_seconds REAL",
-                ),
-                ("file_size_mb", "ALTER TABLE uploads ADD COLUMN file_size_mb REAL"),
-                (
-                    "status",
-                    "ALTER TABLE uploads ADD COLUMN status TEXT DEFAULT 'completed'",
-                ),
-                ("error_message", "ALTER TABLE uploads ADD COLUMN error_message TEXT"),
-                (
-                    "thumbnail_uploaded",
-                    "ALTER TABLE uploads ADD COLUMN thumbnail_uploaded BOOLEAN DEFAULT FALSE",
-                ),
-                (
-                    "ai_analysis_used",
-                    "ALTER TABLE uploads ADD COLUMN ai_analysis_used BOOLEAN DEFAULT FALSE",
-                ),
-                ("created_at", "ALTER TABLE uploads ADD COLUMN created_at DATETIME"),
-                ("updated_at", "ALTER TABLE uploads ADD COLUMN updated_at DATETIME"),
-                # A/B Testing columns
-                (
-                    "thumbnail_ctr_a",
-                    "ALTER TABLE uploads ADD COLUMN thumbnail_ctr_a REAL",
-                ),
-                (
-                    "thumbnail_ctr_b",
-                    "ALTER TABLE uploads ADD COLUMN thumbnail_ctr_b REAL",
-                ),
-                (
-                    "active_thumbnail",
-                    "ALTER TABLE uploads ADD COLUMN active_thumbnail TEXT DEFAULT 'A'",
-                ),
-                (
-                    "thumbnail_test_start_date",
-                    "ALTER TABLE uploads ADD COLUMN thumbnail_test_start_date DATETIME",
-                ),
-                (
-                    "thumbnail_test_complete",
-                    "ALTER TABLE uploads ADD COLUMN thumbnail_test_complete BOOLEAN DEFAULT FALSE",
-                ),
-                (
-                    "winning_thumbnail",
-                    "ALTER TABLE uploads ADD COLUMN winning_thumbnail TEXT",
-                ),
-            ]
-
-            for column_name, migration_sql in migrations:
-                if column_name not in columns:
-                    try:
-                        cursor.execute(migration_sql)
-                        self.logger.info(
-                            f"Database migration: Added column {column_name}"
-                        )
-                    except sqlite3.Error as e:
-                        self.logger.warning(f"Migration failed for {column_name}: {e}")
-
-            # Local artifacts table migration support
-            cursor.execute(
-                "SELECT name FROM sqlite_master WHERE type='table' AND name='local_artifacts'"
-            )
-            artifacts_table_exists = cursor.fetchone() is not None
-
-            if not artifacts_table_exists:
-                self._ensure_local_artifacts_table(cursor)
-            else:
-                cursor.execute("PRAGMA table_info(local_artifacts)")
-                artifact_columns = {row[1] for row in cursor.fetchall()}
-                artifact_migrations = [
-                    (
-                        "local_project_id",
-                        "ALTER TABLE local_artifacts ADD COLUMN local_project_id TEXT",
-                    ),
-                    (
-                        "local_video_path",
-                        "ALTER TABLE local_artifacts ADD COLUMN local_video_path TEXT",
-                    ),
-                    (
-                        "local_thumbnail_paths_json",
-                        "ALTER TABLE local_artifacts ADD COLUMN local_thumbnail_paths_json TEXT",
-                    ),
-                    (
-                        "youtube_video_id",
-                        "ALTER TABLE local_artifacts ADD COLUMN youtube_video_id TEXT",
-                    ),
-                    (
-                        "created_at",
-                        "ALTER TABLE local_artifacts ADD COLUMN created_at DATETIME",
-                    ),
-                    (
-                        "updated_at",
-                        "ALTER TABLE local_artifacts ADD COLUMN updated_at DATETIME",
-                    ),
-                ]
-                for column_name, migration_sql in artifact_migrations:
-                    if column_name not in artifact_columns:
-                        try:
-                            cursor.execute(migration_sql)
-                            self.logger.info(
-                                f"Database migration: Added local_artifacts column {column_name}"
-                            )
-                        except sqlite3.Error as e:
-                            self.logger.warning(
-                                f"Migration failed for local_artifacts.{column_name}: {e}"
-                            )
-
-            cursor.execute(
-                "SELECT name FROM sqlite_master WHERE type='table' AND name='processing_history'"
-            )
-            if cursor.fetchone() is None:
-                self._ensure_processing_history_table(cursor)
-
-        except sqlite3.Error as e:
-            self.logger.warning(f"Error during database migration: {e}")
 
     def _ensure_local_artifacts_table(self, cursor: sqlite3.Cursor) -> None:
         """Create local_artifacts with the canonical schema when missing."""
@@ -565,93 +422,6 @@ class DatabaseManager:
         except sqlite3.Error as e:
             self.logger.error(f"Error getting upload history: {e}")
             return []
-
-    def get_processing_stats(self, days: int = 30) -> Dict[str, Any]:
-        """Get processing statistics for the last N days"""
-        try:
-            days_int = max(0, int(days))
-            days_modifier = f"-{days_int} days"
-            with self.get_connection() as conn:
-                cursor = conn.cursor()
-
-                # Get basic stats - SQL injection vulnerability fixed by parameterizing days_modifier
-                cursor.execute(
-                    """
-                    SELECT 
-                        COUNT(*) as total_uploads,
-                        COUNT(CASE WHEN status = 'completed' THEN 1 END) as successful_uploads,
-                        COUNT(CASE WHEN status = 'failed' THEN 1 END) as failed_uploads,
-                        AVG(processing_duration_seconds) as avg_processing_time,
-                        AVG(video_duration_seconds) as avg_video_duration,
-                        SUM(file_size_mb) as total_file_size_mb
-                    FROM uploads 
-                    WHERE upload_timestamp >= datetime('now', ?)
-                """,
-                    (days_modifier,),
-                )
-
-                stats = dict(cursor.fetchone())
-
-                # Calculate success rate
-                if stats["total_uploads"] > 0:
-                    stats["success_rate"] = (
-                        stats["successful_uploads"] / stats["total_uploads"]
-                    ) * 100
-                else:
-                    stats["success_rate"] = 0
-
-                # Get subreddit breakdown
-                cursor.execute(
-                    """
-                    SELECT subreddit, COUNT(*) as count
-                    FROM uploads 
-                    WHERE upload_timestamp >= datetime('now', ?)
-                    GROUP BY subreddit
-                    ORDER BY count DESC
-                    LIMIT 10
-                """,
-                    (days_modifier,),
-                )
-
-                stats["top_subreddits"] = [dict(row) for row in cursor.fetchall()]
-
-                return stats
-
-        except sqlite3.Error as e:
-            self.logger.error(f"Error getting processing stats: {e}")
-            return {}
-
-    def cleanup_old_records(self, days_to_keep: int = 90) -> Dict[str, int]:
-        """Clean up old records to prevent database bloat"""
-        try:
-            days_int = max(0, int(days_to_keep))
-            days_modifier = f"-{days_int} days"
-            with self.get_connection() as conn:
-                cursor = conn.cursor()
-
-                cursor.execute(
-                    "DELETE FROM uploads WHERE upload_timestamp < datetime('now', ?)",
-                    (days_modifier,),
-                )
-                uploads_deleted = max(0, cursor.rowcount)
-
-                cursor.execute(
-                    "DELETE FROM processing_history WHERE created_at < datetime('now', ?)",
-                    (days_modifier,),
-                )
-                history_deleted = max(0, cursor.rowcount)
-
-                conn.commit()
-                deleted = {
-                    "uploads_deleted": uploads_deleted,
-                    "processing_history_deleted": history_deleted,
-                }
-                self.logger.info("Cleaned old records: %s", deleted)
-                return deleted
-
-        except sqlite3.Error as e:
-            self.logger.error(f"Error cleaning up old records: {e}")
-            return {"uploads_deleted": 0, "processing_history_deleted": 0}
 
     def get_failed_uploads(self, limit: int = 50) -> List[Dict[str, Any]]:
         """Get failed uploads for retry analysis"""
