+++ src/integrations/reddit_client.py
@@ -134,192 +134,6 @@ class RedditPost:
         )
 
 
-class ContentFilter:
-    """Content filtering and validation for Reddit posts"""
-
-    def __init__(self):
-        self.config = get_config()
-        self.logger = logging.getLogger(__name__)
-        self._pattern_cache: Dict[str, Pattern[str]] = {}
-
-    def _build_pattern(self, term: str) -> Pattern[str]:
-        pattern = self._pattern_cache.get(term)
-        if pattern is None:
-            pattern = re.compile(rf"\b{re.escape(term)}\b", re.IGNORECASE)
-            self._pattern_cache[term] = pattern
-        return pattern
-
-    def _find_matching_terms(self, text: Optional[str], terms: List[str]) -> List[str]:
-        if not text or not terms:
-            return []
-
-        matches = [term for term in terms if self._build_pattern(term).search(text)]
-        return sorted(set(matches))
-
-    def _tier_terms(self) -> Dict[str, List[str]]:
-        content_config = self.config.content
-        hard_terms = set(content_config.hard_disallowed)
-        hard_terms.update(content_config.forbidden_words)
-
-        demonetization_terms = set(content_config.demonetization_risk)
-        caution_terms = set(content_config.caution)
-
-        # Backward compatibility for existing configs that only define the old list.
-        demonetization_terms.update(content_config.unsuitable_content_types)
-
-        return {
-            "hard_disallowed": sorted(hard_terms),
-            "demonetization_risk": sorted(demonetization_terms),
-            "caution": sorted(caution_terms),
-        }
-
-    def contains_forbidden_words(self, text: Optional[str]) -> bool:
-        """Check if text contains forbidden words"""
-        terms = self._tier_terms()["hard_disallowed"]
-        return bool(self._find_matching_terms(text, terms))
-
-    def contains_unsuitable_content(self, text: Optional[str]) -> bool:
-        """Check if text contains unsuitable content types"""
-        terms = self._tier_terms()["demonetization_risk"]
-        return bool(self._find_matching_terms(text, terms))
-
-    def is_suitable_for_monetization(self, post: RedditPost) -> Dict[str, Any]:
-        """
-        Comprehensive content suitability check for monetization, including quality.
-
-        Returns:
-            Dict with 'is_suitable', 'reason', and 'confidence' keys
-        """
-        blocking_reasons: List[str] = []
-        flagged_reasons: List[str] = []
-        caution_reasons: List[str] = []
-        quality_reasons: List[str] = []
-
-        tier_terms = self._tier_terms()
-
-        text_fields = {
-            "title": post.title,
-            "subreddit": post.subreddit,
-            "selftext": post.selftext,
-        }
-
-        for field_name, value in text_fields.items():
-            hard_matches = self._find_matching_terms(
-                value, tier_terms["hard_disallowed"]
-            )
-            if hard_matches:
-                blocking_reasons.append(
-                    f"hard_disallowed match in {field_name}: {', '.join(hard_matches)}"
-                )
-
-            demonetization_matches = self._find_matching_terms(
-                value, tier_terms["demonetization_risk"]
-            )
-            if demonetization_matches:
-                flagged_reasons.append(
-                    "demonetization_risk match in "
-                    f"{field_name}: {', '.join(demonetization_matches)}"
-                )
-
-            caution_matches = self._find_matching_terms(value, tier_terms["caution"])
-            if caution_matches:
-                caution_reasons.append(
-                    f"caution match in {field_name}: {', '.join(caution_matches)}"
-                )
-
-        # Check NSFW/spoiler flags
-        if post.nsfw:
-            blocking_reasons.append("NSFW content")
-
-        if post.spoiler:
-            blocking_reasons.append("Spoiler content")
-
-        # Quality checks
-        if post.score < 10:
-            quality_reasons.append("Low score (potential low quality)")
-
-        if post.upvote_ratio < 0.7:
-            quality_reasons.append("Low upvote ratio (controversial content)")
-
-        # Video-specific checks
-        if post.is_video:
-            if post.duration:
-                if post.duration < 5:
-                    quality_reasons.append("Video too short")
-                elif post.duration > 300:  # 5 minutes
-                    quality_reasons.append("Video too long")
-
-            # Source Quality Checks
-            if post.width > 0 and post.height > 0:
-                min_resolution = min(post.width, post.height)
-                if (
-                    min_resolution < 720
-                ):  # Minimum acceptable resolution for Shorts (720p vertical)
-                    quality_reasons.append(
-                        f"Low source resolution ({post.width}x{post.height})"
-                    )
-
-            if post.fps > 0 and post.fps < 25:  # Minimum acceptable FPS
-                quality_reasons.append(f"Low source FPS ({post.fps})")
-
-        is_suitable = len(blocking_reasons) == 0 and len(quality_reasons) == 0
-        total_issues = (
-            len(blocking_reasons)
-            + len(quality_reasons)
-            + len(flagged_reasons)
-            + len(caution_reasons)
-        )
-        confidence = 95 if is_suitable else max(60, 100 - total_issues * 10)
-
-        reason_parts = []
-        if blocking_reasons:
-            reason_parts.append("blocked: " + "; ".join(blocking_reasons))
-        if quality_reasons:
-            reason_parts.append("quality: " + "; ".join(quality_reasons))
-        if flagged_reasons:
-            reason_parts.append("flagged: " + "; ".join(flagged_reasons))
-        if caution_reasons:
-            reason_parts.append("caution: " + "; ".join(caution_reasons))
-
-        return {
-            "is_suitable": is_suitable,
-            "reason": "; ".join(reason_parts)
-            if reason_parts
-            else "Content appears suitable",
-            "confidence": confidence,
-            "blocking_reasons": blocking_reasons,
-            "quality_reasons": quality_reasons,
-            "demonetization_risk": flagged_reasons,
-            "caution_flags": caution_reasons,
-        }
-
-    def filter_posts(self, posts: List[RedditPost]) -> List[RedditPost]:
-        """Filter posts for suitability"""
-        suitable_posts = []
-
-        for post in posts:
-            suitability = self.is_suitable_for_monetization(post)
-
-            if suitability["is_suitable"]:
-                suitable_posts.append(post)
-                if suitability["demonetization_risk"] or suitability["caution_flags"]:
-                    self.logger.info(
-                        "Post accepted with flags: %s... | %s",
-                        post.title[:50],
-                        suitability["reason"],
-                    )
-                else:
-                    self.logger.debug(f"Post accepted: {post.title[:50]}...")
-            else:
-                self.logger.warning(
-                    "Post rejected: %s... | %s",
-                    post.title[:50],
-                    suitability["reason"],
-                )
-
-        return suitable_posts
-
-
 class RedditClient:
     """Async Reddit API client for fetching and managing posts"""
 
@@ -327,7 +141,7 @@ class RedditClient:
         self.config = get_config()
         self.logger = logging.getLogger(__name__)
         self.reddit: Optional[asyncpraw.Reddit] = None
-        self.content_filter = ContentFilter()
+
         self._initialized = False
 
     async def __aenter__(self):
@@ -483,109 +297,6 @@ class RedditClient:
             self.logger.error(f"Error fetching posts from r/{subreddit_name}: {e}")
             return []
 
-    async def fetch_posts_from_multiple_subreddits(
-        self, subreddit_names: Optional[List[str]] = None, posts_per_subreddit: int = 5
-    ) -> List[RedditPost]:
-        """
-        Fetch posts from multiple subreddits
-
-        Args:
-            subreddit_names: List of subreddit names (uses curated list if None)
-            posts_per_subreddit: Number of posts to fetch from each subreddit
-        """
-        if subreddit_names is None:
-            subreddit_names = self.config.content.curated_subreddits
-
-        all_posts = []
-
-        for subreddit_name in subreddit_names:
-            posts = await self.fetch_posts_from_subreddit(
-                subreddit_name, limit=posts_per_subreddit
-            )
-            all_posts.extend(posts)
-
-        self.logger.info(
-            f"Fetched {len(all_posts)} total posts from {len(subreddit_names)} subreddits"
-        )
-        return all_posts
-
-    async def get_filtered_video_posts(
-        self,
-        subreddit_names: Optional[List[str]] = None,
-        max_posts: Optional[int] = None,
-    ) -> List[RedditPost]:
-        """
-        Get filtered video posts suitable for processing
-
-        Args:
-            subreddit_names: List of subreddit names to search
-            max_posts: Maximum number of posts to return
-
-        Returns:
-            List of suitable video posts
-        """
-        if max_posts is None:
-            max_posts = self.config.content.max_reddit_posts_to_fetch
-
-        # Fetch posts from multiple subreddits
-        all_posts = await self.fetch_posts_from_multiple_subreddits(
-            subreddit_names,
-            posts_per_subreddit=max(
-                5,
-                max_posts
-                // len(subreddit_names or self.config.content.curated_subreddits),
-            ),
-        )
-
-        # Filter for video posts only
-        video_posts = [post for post in all_posts if post.is_video and post.video_url]
-        self.logger.info(
-            f"Found {len(video_posts)} video posts out of {len(all_posts)} total posts"
-        )
-
-        # Apply content filtering
-        suitable_posts = self.content_filter.filter_posts(video_posts)
-        self.logger.info(f"Filtered to {len(suitable_posts)} suitable posts")
-
-        # Sort by score and return top posts
-        suitable_posts.sort(key=lambda p: p.score, reverse=True)
-        result = suitable_posts[:max_posts]
-
-        self.logger.info(f"Returning top {len(result)} posts for processing")
-        return result
-
-    async def get_lore_text_posts(
-        self, subreddit_names: Optional[List[str]] = None, max_posts: int = 5
-    ) -> List[RedditPost]:
-        """Fetch highly upvoted long-form text posts for lore script generation."""
-        if subreddit_names is None:
-            subreddit_names = self.config.content.curated_subreddits
-
-        all_posts = await self.fetch_posts_from_multiple_subreddits(
-            subreddit_names, posts_per_subreddit=10
-        )
-
-        lore_posts = [
-            post
-            for post in all_posts
-            if (
-                not post.is_video
-                and isinstance(post.selftext, str)
-                and len(post.selftext.strip()) > 300
-            )
-        ]
-
-        suitable_posts = self.content_filter.filter_posts(lore_posts)
-        suitable_posts.sort(key=lambda p: p.score, reverse=True)
-        selected = suitable_posts[:max_posts]
-
-        self.logger.info(
-            "Selected %d lore text posts from %d total fetched posts",
-            len(selected),
-            len(all_posts),
-        )
-        return selected
-
     async def get_post_by_url(self, url: str) -> Optional[RedditPost]:
         """Get a specific post by Reddit URL"""
         if not self.is_connected():
@@ -632,85 +343,3 @@ class RedditClient:
         except Exception as e:
             self.logger.error(f"Error fetching post from URL {url}: {e}")
             return None
-
-    async def get_post_by_url_async(self, url: str) -> Optional[RedditPost]:
-        """Get a specific post by Reddit URL using async context manager"""
-        async with self:
-            return await self.get_post_by_url(url)
-
-    async def search_posts(
-        self,
-        query: str,
-        subreddit_names: Optional[List[str]] = None,
-        sort: str = "relevance",
-        time_filter: str = "week",
-        limit: int = 10,
-    ) -> List[RedditPost]:
-        """
-        Search for posts matching a query
-
-        Args:
-            query: Search query
-            subreddit_names: List of subreddits to search (all if None)
-            sort: 'relevance', 'hot', 'top', 'new', 'comments'
-            time_filter: Time filter for search
-            limit: Maximum results to return
-        """
-        if not self.is_connected():
-            await self._initialize_client()
-            if not self.is_connected():
-                return []
-
-        try:
-            if subreddit_names:
-                # Search in specific subreddits
-                subreddit_str = "+".join(subreddit_names)
-                subreddit = await self.reddit.subreddit(subreddit_str)
-            else:
-                # Search all of Reddit
-                subreddit = await self.reddit.subreddit("all")
-
-            submissions = subreddit.search(
-                query=query, sort=sort, time_filter=time_filter, limit=limit
-            )
-
-            posts = []
-            async for submission in submissions:
-                try:
-                    post = RedditPost.from_submission(submission)
-                    posts.append(post)
-                except Exception as e:
-                    self.logger.warning(
-                        f"Error processing search result {submission.id}: {e}"
-                    )
-                    continue
-
-            self.logger.info(f"Found {len(posts)} posts matching query: {query}")
-            return posts
-
-        except asyncprawcore.ResponseException as e:
-            if e.response.status == 401:
-                self.logger.error(
-                    "Authentication failed during search: Reddit API credentials invalid or expired"
-                )
-                self.logger.error(
-                    "Please check your REDDIT_CLIENT_ID and REDDIT_CLIENT_SECRET in the .env file"
-                )
-                self.logger.error(
-                    "Run 'python test_reddit_connection.py' to verify your Reddit API setup"
-                )
-            else:
-                self.logger.error(
-                    f"Reddit API error during search: HTTP {e.response.status} - {e}"
-                )
-            return []
-        except Exception as e:
-            self.logger.error(f"Error searching posts: {e}")
-            return []
-
-
-async def create_reddit_client() -> RedditClient:
-    """Factory function to create an async Reddit client"""
-    client = RedditClient()
-    await client._initialize_client()
-    return client
